requirements.txt
flask
pytest
gunicorn
pyyaml